requirements.txt
langchain
langchain-openai
langchain-community
beautifulsoup4
trafilatura
faiss-cpu
python-dotenv
requests